backend/requirements.txt
flask
pillow
qrcode
python-dotenv
bcrypt
PyJWT
flask-cors
datetime
